Add an end-to-end test for the first scenario
In order to actually explore the boundaries of our scenario, such as the UI interaction, we write a test that runs on a real device.
This test will exercise the scenario code and the android frameworks, and will remain red until we have implemented the required UI elements.
Thus, in our presentation package we are forced to deal with android specific concepts like activities, adapters and managed lifecycle.
We decide to use Robotium as our IT library as it make easier to write expressive and more robust tests, as independent as possible from the actual implementation.