Parse a search repository response into Java models
We decide to use Jackson as a Json parsing library for its performance, flexibility and annotation-based configuration.
To ensure maximum flexibility, we hide the ObjectMapper object with our own implementation, JsonConverter.
We use custom converters for every entity, so that the object creation must pass through the constructor.
This way our models remain immutable and we can enforce class invariants such as "startDate < endDate".
This also keeps Jackson as an implementation detail and we're free to swap for another library (like Gson) without touching our entities.
We make sure that entities that are Collections are treated as first class objects, because they represent a concept in our domain and may contain specific business logic.