Write a stub for the Github api
We write an implementation for the GithubApi interface that will provide predictable responses for every query.
That way we can run the entire application without having to rely on the Github servers, speeding up both testing and development.
Also, those mocked responses allow us to eliminate some duplication in the test classes.