Add a ScenarioDispatcher as a convenience central point to execute scenarios
The Dispatcher is the class that the boundaries will use to access the ability to run scenarios.
This is a major abstraction on how the scenarios are effectively executed: they might be run directly, put into a queue or scheduled for later, but the boundaries should not care about it.
We also choose to assign to the ScenarioFactory the responsibility for wrapping the actual scenario and the presentation on separate threads.
That might be an odd solution, but it works for now and we can refactor later if the need arieses.