Bridge the gap between the presentation layer and the android lifecycle.
Maintaining references to objects in Android can be dangerous because the platform may destroy them and reclaim the memory at any time.
We need a solution to decouple our business logic from the details of the android framework.
The two widely adopted solutions are an Event Bus and an Injection Framework.
The bus allow us to broadcast a certain event without holding an explicit reference to the objects that will respond to this event.
Having an event bus in place means that we can cache and manipulate asynchronous responses easier than if we used AsyncTasks or Loaders.
Injection allow us to add dependencies to the object that we can't construct, like activities and fragments, makes our architecture cleaner and with less boilerplate code and makes writing tests easier.