Add test fixtures to the real Github service
Our integration tests are green, but we are running those tests against a mock server.
In order to test that our predictions are correct and that the apis are stable, we can exercise our code against the live service.
This step can be taken in parallel with the development of the rest of the application code.
Since we are exploring another boundary, the api service, we wrap the implementation details with our own interface, the GithubApi.
We pick Retrofit as a REST framework for its ability to run both on Android and on JVM.