Define a Controller to drive the activity behavior
The easiest way to TDD an activity is to write a controller for it.
You'll start to test the behavior that you want to accomplish with that component, like handling a search query or update the ActionBar information.
Notice that there are no notions of intents or instanceState or lifecycle callbacks in our controller.
That's a good way to separate how a certain behavior is accomplished from the behavior itself.